[FIX] 경매, 입찰 시간 충돌방지 수정
